🎨 画像生成部分のコンポーネント化
